fix migration

+++ db/migrate/20180818131758_create_event_schedules.rb
@@ -1,7 +1,7 @@
 class CreateEventSchedules < ActiveRecord::Migration[5.2]
   def change
     create_table :event_schedules do |t|
-      t.integer :event_id, null: false
+      t.bigint :event_id, null: false
       t.datetime :started_at, null: false
       t.datetime :ended_at, null: false
       t.string :address
